Proposed abstraction of patch building tools into a PatchBuilder class to be shared between the PatchTableFactory and new topology hashing algorithms.
- add PatchBuilder class
- inherit PatchBuilderContext from PatchBuilder in PatchTableFactory
- move end cap types enum to types.h
- propagate 'fallout' from these changes
